+++ src/Application/Requests/Events/AbstractEventRequest.php
@@ -24,7 +24,7 @@ abstract class AbstractEventRequest extends AbstractRequest implements EventInte
 
     protected array $eventPayload;
 
-    protected int $eventId;
+    protected ?int $eventId;
 
     public function __construct(Request $request)
     {
@@ -32,13 +32,12 @@ abstract class AbstractEventRequest extends AbstractRequest implements EventInte
         $payload = [];
         parse_str($request->getContent(), $payload);
         $this->eventPayload = $payload;
-
         $this->eventCode = $this->eventPayload['event'];
         $this->timestamp = (int)$this->eventPayload['ts'];
-        $this->eventId = (int)$this->eventPayload['event_id'];
+        $this->eventId = array_key_exists('event_id', $this->eventPayload) ? (int)$this->eventPayload['event_id'] : null;
     }
 
-    public function getEventId(): int
+    public function getEventId(): ?int
     {
         return $this->eventId;
     }