+++ src/Application/Requests/Events/OnApplicationUninstall/OnApplicationUninstall.php
@@ -17,6 +17,8 @@ use Bitrix24\SDK\Application\Requests\Events\AbstractEventRequest;
 
 class OnApplicationUninstall extends AbstractEventRequest
 {
+    public const CODE = 'ONAPPUNINSTALL';
+
     public function getApplicationData(): ApplicationData
     {
         return new ApplicationData($this->eventPayload['data']);